% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracyScores.R
\name{accuracyScores}
\alias{accuracyScores}
\title{Compute accuracy scores.}
\usage{
accuracyScores(data = NA, type = "local", method = "correlation")
}
\arguments{
\item{data}{data.}

\item{type}{string: what kind of scores should be computed ("local", default,
or "global"). If set to "global", the \code{igraph} library needs to be
available.}

\item{method}{string: what method to use in order to compute scores
("correlation", default, or "cosine", "ejaccard", "edice",
 "bhjattacharyya", "canberra", "divergence", "wave", "kullback",
 "bray", "soergel", "chord", "geodesic", "whittaker", "hellinger").
Except from "correlation", the other methods require the \code{proxy}
library.}
}
\value{
Reshaped data.
}
\description{
Compute accuracy scores.
}
