% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptTradeDataNames.R
\name{adaptTradeDataNames}
\alias{adaptTradeDataNames}
\title{Standardise TL or ES variable names.}
\usage{
adaptTradeDataNames(tradedata, origin)
}
\arguments{
\item{tradedata}{TL or ES trade data.}

\item{origin}{String: "TL" or "ES", for Tariff Line or Eurostat data
respectively.}
}
\value{
TL or ES data with common names (TL will also have "qunit").
}
\description{
Use the same variable names in both datasets.
}
\details{
TL and ES data use different variable names for reporters, partners,
commodities, values, quantity, and year. This function give these
variables the same name in both datasets.
}
