% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptTradeDataTypes.R
\name{adaptTradeDataTypes}
\alias{adaptTradeDataTypes}
\title{Standardise TL or ES variable names.}
\usage{
adaptTradeDataTypes(tradedata, origin)
}
\arguments{
\item{tradedata}{TL or ES trade data.}

\item{origin}{String: "TL" or "ES", for Tariff Line or Eurostat data
respectively.}
}
\value{
TL or ES data with common data types.
}
\description{
Use the same variable types in both datasets: integer for year, reporter,
  partner, flow; numeric for value, weight and qty; character for hs and
  created hs6.
}
