% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cpcs_from_links.R
\name{add_cpcs_from_links}
\alias{add_cpcs_from_links}
\title{Add cpc codes to trade data from links datasets}
\usage{
add_cpcs_from_links(tradedata, hs6links, links)
}
\arguments{
\item{tradedata}{Data.}

\item{hs6links}{HS6 links.}

\item{links}{HS-CPC links.}
}
\description{
Add cpc codes to trade data from links datasets
}
