% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fcls_from_links.R
\name{add_fcls_from_links}
\alias{add_fcls_from_links}
\title{Add fcl codes to trade data from links datasets}
\usage{
add_fcls_from_links(tradedata, hs6links, links)
}
\description{
Add fcl codes to trade data from links datasets
}
