% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFCLunits.R
\name{addFCLunits}
\alias{addFCLunits}
\title{Add FCL variable to trade data.}
\usage{
addFCLunits(tradedata = NA, fclunits = NA)
}
\arguments{
\item{tradedata}{Trade data.}

\item{fclunits}{A table containing FCL codes and expected measure units.}
}
\value{
\code{tradedata} with \code{fclunits} added.
}
\description{
For each FCL code an expected FCL unit (i.e., what units does FAO expects
for given FCL codes) is added.
}
