% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFinalUnits.R
\name{addFinalUnits}
\alias{addFinalUnits}
\title{Add "finalunits" variable to trade data.}
\usage{
addFinalUnits(tradedata = NA, finalunits = NA)
}
\arguments{
\item{tradedata}{Trade data.}

\item{finalunits}{A table containing CPC codes and expected measure units.}
}
\value{
\code{tradedata} with \code{finalunits} added.
}
\description{
For each CPC code an expected CPC unit (i.e., what units does FAO expects
for given CPC codes) is added.
}
