% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addRetradeToTrade}
\alias{addRetradeToTrade}
\title{This function adds retrade to trade for both quantity and value.}
\usage{
addRetradeToTrade(data, importQuantity, reimportQuantity, exportQuantity,
  reexportQuantity, importValue, reimportValue, exportValue, reexportValue)
}
\arguments{
\item{data}{The data frame}

\item{importQuantity}{The column representing the quantity of import.}

\item{reimportQuantity}{The column representing the quantity of re-import.}

\item{exportQuantity}{The column representing the quantity of export.}

\item{reexportQuantity}{The column representing the quantity of re-export.}

\item{importValue}{The column representing the value of import.}

\item{reimportValue}{The column representing the value of re-import.}

\item{exportValue}{The column representing the value of export.}

\item{reexportValue}{The column representing the value of re-export.}
}
\description{
This function adds retrade to trade for both quantity and value.
}

