% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignHSconvert2FCL.R
\name{alignHSconvert2FCL}
\alias{alignHSconvert2FCL}
\title{Align HS codes in trade data set and in HS2FCL map and converts them to FCL}
\usage{
alignHSconvert2FCL(tldata, hsfclmap, parallel)
}
\description{
Align HS codes in trade data set and in HS2FCL map and converts them to FCL
}

