% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignHSLength.R
\name{alignHSLength}
\alias{alignHSLength}
\title{Prepares for numerical comparison vector of hs codes from 
trade dataset with hs codes range from mapping table.}
\usage{
alignHSLength(hs, mapdataset)
}
\arguments{
\item{hs}{Vector of original hs codes from trade data set.}

\item{mapdataset}{Data frame with mapping table containing 
at least columns area, flow, fromcode, tocode}
}
\value{
A list with two components: 
  * hs - vector with extended hs codes
  * mapdataset - data frame with extended fromcode and to code
  columns.
}
\description{
Prepares for numerical comparison vector of hs codes from 
trade dataset with hs codes range from mapping table.
}
\details{
Alignes length of hs codes in three places: vector from trade 
  dataset, fromcode and tocode columns of mapping dataset. 
  Maximum length is determined and all shorter codes are extended 
  on right-hand side: trade data hs code and mapping 
  fromcode are extended by 0, mapping tocode is extended by 9.
}

