% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyadj.R
\name{applyadj}
\alias{applyadj}
\title{Apply adjustments to trade data}
\usage{
applyadj(rep, yr, adjustments, tradedata, dbg = FALSE)
}
\arguments{
\item{rep}{Reporter.}

\item{yr}{Year.}

\item{adjustments}{Adjustments table.}

\item{tradedata}{Trade data.}

\item{dbg}{Debug option (TRUE or FALSE).}
}
\description{
Apply adjustments to trade data
}
