% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_name_fao.R
\name{area_name_fao}
\alias{area_name_fao}
\title{A function to provide names for FAO area list codes.}
\usage{
area_name_fao(areacode)
}
\arguments{
\item{areacode}{Vector of FAO area list codes.}
}
\value{
Character vector of area names.
}
\description{
A function to provide names for FAO area list codes.
}
