% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_utilities.R
\name{bind_rprts}
\alias{bind_rprts}
\title{Bind similar data frames from different years}
\usage{
bind_rprts(rprt_elems_list)
}
\arguments{
\item{rprt_elems_list}{List produced by `extract_rprt_elem`.}
}
\value{
Data frame if there is one report, and list of frames if more.
}
\description{
Bind similar data frames from different years
}
