% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateReliability}
\alias{calculateReliability}
\title{This function calculates the trade reliability for the reporting
country based on quantity.}
\usage{
calculateReliability(data, import, export, reverseImport, reverseExport,
  reportingCountry, partnerCountry, pctTolerance)
}
\arguments{
\item{data}{The data table.}

\item{import}{The import quantity reported by the reporting country.}

\item{export}{The export quantity reported by the reporting country.}

\item{reversemport}{The import quantity reported by the partner country.}

\item{reverseExport}{The export quantity reported by the partner country.}

\item{reportingCountry}{The column representing the reporting country.}

\item{partnerCountry}{The column representing the partner country.}

\item{pctTolerance}{Percentage, the tolerance level of discrepancy
between trading partner is deemd tolerable.}
}
\description{
This function calculates the trade reliability for the reporting
country based on quantity.
}

