% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateUnitValue}
\alias{calculateUnitValue}
\title{This function calculates the unit value of the trade.}
\usage{
calculateUnitValue(data, importUnitValue, importTradeValue, importTradeQuantity,
  exportUnitValue, exportTradeValue, exportTradeQuantity)
}
\arguments{
\item{data}{The data frame}

\item{importUnitValue}{The column representing the import unit value.}

\item{importTradeValue}{The column representing the import trade value.}

\item{importTradeQuantity}{The column representing the import trade quantity.}

\item{exportUnitValue}{The column representing the export unit value.}

\item{exportTradeValue}{The column representing the export trade value.}

\item{exportTradeQuantity}{The column representing the export trade quantity.}
}
\description{
This function calculates the unit value of the trade.
}

