% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_versions.R
\name{check_versions}
\alias{check_versions}
\title{Check package versions requirements}
\usage{
check_versions(packages, versions)
}
\arguments{
\item{packages}{Character vector of packages' names to check}

\item{versions}{Character vectorr of minimal packages' versions}
}
\description{
Check package versions requirements
}
