% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMedianUnitValue.R
\name{computeMedianUnitValue}
\alias{computeMedianUnitValue}
\title{Compute the median unit value of the transaction.}
\usage{
computeMedianUnitValue(tradedata = NA, name = "uvm", other = TRUE)
}
\arguments{
\item{tradedata}{Trade data.}

\item{other}{Logical value indicating whether other median unit values
and the number of reporters are reported. By default it is TRUE.}

\item{output}{A string specifying the name of the variable name that
will contain the median unit value ("uvm" by default).}
}
\value{
\code{tradedata} with median unit value. If \code{other} == TRUE
  intermediate median unit values and the number of partners on which the
  medians are computed are returned.
}
\description{
The median unit value is calculated in a specific-to-generic fashion:
1) by HS-reporter; 2) by HS8-reporter; by HS6-reporter; 3) by HS;
4) by FCL; 5) by flow.
The first available of these options will be used (for HS, HS8 and HS6
only when there are at least 10 flows).
}
