% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comtradeunits}
\alias{comtradeunits}
\title{Units of measuremnt for Comtrade datasets}
\format{A data frame: qunit (integer) UN Comtrade Code,
wco (character) - WCO Abbreviation,
desc (character) - Description}
\usage{
comtradeunits
}
\description{
Taken from Annex I of
http://unstats.un.org/unsd/tradekb/Knowledgebase/Quantity-and-Weight-Data-in-UN-Comtrade
}
\keyword{datasets}
