% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertComtradeM49ToFAO.R
\name{convertComtradeM49ToFAO}
\alias{convertComtradeM49ToFAO}
\title{Converts area codes, stored in UNSD Comtrade M49, into FAO area codes}
\usage{
convertComtradeM49ToFAO(reporters)
}
\arguments{
\item{Numeric}{or character vector of UNSC M49 area codes}
}
\value{
Interger vector of FAO area codes
}
\description{
Converts area codes, stored in UNSD Comtrade M49, into FAO area codes
}

