% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeonom2FAO.R
\name{convertGeonom2FAO}
\alias{convertGeonom2FAO}
\title{Convertion of EuroStat Geonomenclature to FAO area codes}
\usage{
convertGeonom2FAO(geonom)
}
\description{
Convertion of EuroStat Geonomenclature to FAO area codes
}

