% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHS2FCL.R
\name{convertHS2FCL}
\alias{convertHS2FCL}
\title{Converts dataset with tradeflows from HS to FCL}
\usage{
convertHS2FCL(tradedata, hsfclmap, parallel)
}
\description{
It is based on notes from FAO ESS MDB files.
}
