% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMeasuredElementTrade.R
\name{convertMeasuredElementTrade}
\alias{convertMeasuredElementTrade}
\title{Converts measuredElementTrade for the output}
\usage{
convertMeasuredElementTrade(element, unit, flow)
}
\arguments{
\item{element}{Element.}

\item{unit}{Unit.}

\item{flow}{Flow.}
}
\description{
Converts measuredElementTrade for the output
}
