% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTLPartnerToFAO.R
\name{convertTLPartnerToFAO}
\alias{convertTLPartnerToFAO}
\title{Convert tariffline partner code to FAO code}
\usage{
convertTLPartnerToFAO(partners)
}
\description{
Convert tariffline partner code to FAO code
}

