% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Detect anomalous observations (outliers).}
\usage{
detectOutliers(tradedata = NA, method = NA, parameters = NA)
}
\arguments{
\item{tradedata}{Trade data.}

\item{method}{String: method to use for outlier detection. Currently
"boxplot" is allowed.}

\item{parameters}{List: named parameters to pass to \code{method}.}
}
\value{
\code{tradedata} with an additional "outlier" column: TRUE
  indicates an outlier.
}
\description{
An anomalous unit value will be indicated as "outlier".
}

