% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doImputation.R
\name{doImputation}
\alias{doImputation}
\title{Impute missing values or outliers.}
\usage{
doImputation(tradedata = NA, uv.median = "uvm")
}
\arguments{
\item{tradedata}{Trade data.}

\item{uv.median}{Name of the median unit value variable.}
}
\value{
\code{tradedata} with imputed values.
}
\description{
Use a median unit value to impute missing quantities or outliers.
}
\details{
The median unit value is calculated in a specific-to-generic fashion:
1) by HS-reporter; 2) by HS8-reporter; by HS6-reporter; 3) by HS;
4) by FCL; 5) by flow.
The first available of these options will be used (for HS, HS8 and HS6
only when there are at least 10 flows).
}
