% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hs6cpcmap.R
\name{extract_hs6cpcmap}
\alias{extract_hs6cpcmap}
\title{Create HS6->CPC mapping table.}
\usage{
extract_hs6cpcmap(maptable = NULL)
}
\arguments{
\item{maptable}{hscpcmap data frame.}
}
\value{
Data frame with columns reporter, flow, hs6, cpc
}
\description{
Create HS6->CPC mapping table.
}
