% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hs6fclmap.R
\name{extract_hs6fclmap}
\alias{extract_hs6fclmap}
\title{Create HS6->FCL mapping table.}
\usage{
extract_hs6fclmap(maptable = NULL, parallel = FALSE)
}
\arguments{
\item{maptable}{hsfclmap data frame.}

\item{parallel}{Should parallel execution be used if available. FALSE by
default.}
}
\value{
Data frame with columns reporter, flow, hs6, fcl
}
\description{
Create HS6->FCL mapping table.
}
