% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_utilities.R
\name{extract_rprt_year}
\alias{extract_rprt_year}
\title{Extracts year from report directory name}
\usage{
extract_rprt_year(dirname = NULL, prefix = NULL)
}
\arguments{
\item{dirname}{Character vector with directories' names. NULL by default.}

\item{prefix}{A string with report directory name prefix (without trailing
underscore. By default NULL.}
}
\value{
Integer vector same length as `dirname` with years.
}
\description{
Extracts year from report directory name
}
