% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_country_code}
\alias{fao_country_code}
\title{Information of country code and names}
\format{A data frame with two columns as integer and character:
reporter and country_name}
\usage{
fao_country_code
}
\description{
Taken from data_raw, put by Alex
}
\keyword{datasets}
