% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoAreaName.R
\name{faoAreaName}
\alias{faoAreaName}
\title{Assign area names to M49 or geonom codes.}
\usage{
faoAreaName(areacode, swsdomain = "faostat_one", swsdataset = "FS1_SUA_UPD",
  swsdimension = "geographicAreaFS")
}
\arguments{
\item{areacode}{Area code.}

\item{swsdomain}{SWS domain name (default = "faostat_one").}

\item{swsdataset}{SWS dataset (default = "FS1_SUA_UPD").}

\item{swsdimension}{(default = "geographicAreaFS").}
}
\description{
Assign area names to M49 or geonom codes.
}
