% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fcl_2_cpc2}
\alias{fcl_2_cpc2}
\title{FAO Commodity List (FCL) to Central Product Classification (CPC)
Version 2.1 extended.}
\format{A data frame:
fcl (character) - FAO Commodity List (FCL),
cpc (character) - Central Product Classification (CPC2.1) extended}
\usage{
fcl_2_cpc2
}
\description{
This table was created from a more exhaustive table provided by Giorgio
and present in the data-raw (Conv_3DEC2015_E.XLSX)
}
\keyword{datasets}

