% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fclcodelist}
\alias{fclcodelist}
\title{Local replacement for table with fcl codes and descriptions from SWS}
\format{data.table}
\usage{
fclcodelist
}
\description{
Local replacement for table with fcl codes and descriptions from SWS
}
\keyword{datasets}
