% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fclunits}
\alias{fclunits}
\title{Units of measurment for FCL codes}
\format{A data frame with two columns: integer FCL code
and character description of a unit}
\usage{
fclunits
}
\description{
Taken as XLS from Claudia DeVita Claudia.DeVita@fao.org
}
\keyword{datasets}

