% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterHS6FAOinterest.R
\name{filterHS6FAOinterest}
\alias{filterHS6FAOinterest}
\title{Function to filter trade data by HS6 codes of FAO interest}
\usage{
filterHS6FAOinterest(tradedata)
}
\description{
Fiter out HS codes which don't participate in futher processing
Such solution drops all HS codes shorter than 6 digits.
}

