% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsToMirror.R
\name{flowsToMirror}
\alias{flowsToMirror}
\title{Check which flows for which countries need to be mirrored.}
\usage{
flowsToMirror(data = NA, names = FALSE)
}
\arguments{
\item{data}{Trade dataset with ALL reporters included.}

\item{names}{Logical: if TRUE an additional column is
added to the result with country names (see
\code{\link[faoswsTrade]{faoAreaName}}; if FALSE (default)
such column is not added.}
}
\value{
A table with two columns: \code{area} is the country
  (that could originally have been a reporter or a partner)
  code, and \code{flow} that is an integer representing imports
  (1) or exports (2).
}
\description{
Some countries may be listed as partners, but never show up
as reporters, while other may have an incomplete coverage of
reported flows (e.g., imports are reported, but not exports
and partner report having imports from that country). This
function checks when one of these two cases happen and build
a table of them.
}
