% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFlagVars.R
\name{generateFlagVars}
\alias{generateFlagVars}
\title{Generate flag variables in trade data.}
\usage{
generateFlagVars(data = stop("'data' needs to be set."), status = c("E",
  "I", "X"), method = c("c", "i", "e", "s", "h"))
}
\description{
Generate flag variables in trade data.
}
