% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geonom2fao}
\alias{geonom2fao}
\title{Eurostat Geonomenclature to FAO area codes mapping}
\format{A data frame: code (integer) Geonom code,
faostat (integer) - FAO area code,
active (integer) - FAO area code to use. For example, French Guiana
goes to France. So we don't need to aggregate later (probably). Abandoned codes,
like USSR, in active column were equal 0, but replaced by NA.
name (character) - name of the area code}
\usage{
geonom2fao
}
\description{
Extracted from CountryCode table of FRA_2012.mdb located in
 /mnt/essdata/TradeSys/TradeSys/Countries/FRA_2012/
All EU countries have similar codes. So we took France.
Idea suggested by Claudia
}
\keyword{datasets}
