% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgriHSCodes.R
\name{getAgriHSCodes}
\alias{getAgriHSCodes}
\title{Returns subset of HS-codes of agriculture items used by ESS devision.}
\usage{
getAgriHSCodes()
}
\value{
Character vector
}
\description{
List of agricultural commodities provided by Cladia DeVita.
}

