% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllItems.R
\name{getAllItems}
\alias{getAllItems}
\title{Retrieves data on HS-codes and their descriptions from SWS}
\usage{
getAllItems(dmn = "trade", dtset = "ct_raw_tf",
  dimension = "measuredItemHS")
}
\value{
Data.table with code and description
}
\description{
Retrieves data on HS-codes and their descriptions from SWS
}
