% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComputedDataSWS.R
\name{getComputedDataSWS}
\alias{getComputedDataSWS}
\title{Get computed dataset on SWS.}
\usage{
getComputedDataSWS(reporter = NA, omit = FALSE)
}
\arguments{
\item{reporter}{Reporter.}

\item{omit}{Logical indicating whether to omit or not NAs.}
}
\value{
A data.table with results.
}
\description{
Get computed dataset on SWS.
}
