% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHsManyLengths.R
\name{getHsManyLengths}
\alias{getHsManyLengths}
\title{Get HS codes with many digits.}
\usage{
getHsManyLengths(data)
}
\arguments{
\item{data}{Trade data including both TL and ES.}
}
\description{
This function selects the HS codes that have many lengths.
}
