% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlistofadjs.R
\name{getlistofadjs}
\alias{getlistofadjs}
\title{Generate list of adjustments from a table extracted from MDB notes}
\usage{
getlistofadjs(rep, yr, adjustments)
}
\description{
Generate list of adjustments from a table extracted from MDB notes
}

