% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSaveMostRecentfolders.R
\name{getSaveMostRecentfolders}
\alias{getSaveMostRecentfolders}
\title{Get and save the most updated folders.}
\usage{
getSaveMostRecentfolders(initialDir = initial, saveDir = save,
  minYear = minYearToProcess, maxYear = maxYearToProcess)
}
\arguments{
\item{initialDir}{Path where there are the trade folders.}

\item{saveDir}{Path where the most updated trade folders should be saved.}

\item{minYear}{Minimum year to be processed the Pre-Processing Report.}

\item{maxYear}{Maximum year to be processed the Pre-Processing Report.}
}
\value{
Folders copied to a specific directory.
}
\description{
This function selects the most updated trade folders and copies specific files
to another directory. The files copied are: development.log and the entire folder datadir.
The subfolder "datadir" is essential for creating the Pre-Processing Report.
}
