% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTradeData.R
\name{getTradeData}
\alias{getTradeData}
\title{Get Tariff Line (TL) or Eurostat (ES) raw data.}
\usage{
getTradeData(origin = NA, year = NA, test = FALSE, local = FALSE)
}
\arguments{
\item{origin}{String: "TL" (Tariff Line) or "ES" (Eurostat).}

\item{year}{Numeric.}

\item{test}{Logical: FALSE (default), complete data; TRUE, test data.}

\item{local}{Logical: FALSE (default), download test; TRUE, local test data.}
}
\value{
A data.table containing trade data.
}
\description{
Get Tariff Line (TL) or Eurostat (ES) raw data.
}

