% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6faointerest}
\alias{hs6faointerest}
\title{HS codes of length 6 from Standard 2012 what take part in futher processing
of agricultural trade data.}
\format{A character vector of length 1061.}
\usage{
hs6faointerest
}
\description{
Taken from HS2012-6 digits Standard.xls file provided by Cladia from B/C team.
Version of December 2016. Consists of 1061 codes.
https://github.com/SWS-Methodology/faoswsTrade/issues/66
}
\keyword{datasets}

