% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsfclmapSubset.R
\name{hsfclmapSubset}
\alias{hsfclmapSubset}
\title{Subset the HS to FCL map.}
\usage{
hsfclmapSubset(map = NA, year = NA)
}
\arguments{
\item{map}{HS to FCL map.}

\item{year}{Numeric.}
}
\value{
A tbl which is a subset of \code{map}.
}
\description{
Remove irrelevant years in the HS to FCL mapping. When there is not
a specific year for the mapping, take the nearest year.
}

