% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsInRange.R
\name{hsInRange}
\alias{hsInRange}
\title{Looks for corresponding FCL codes in country-specific
mapping tables from MDB files}
\usage{
hsInRange(hs, areacode, flowname, mapdataset, parallel = F)
}
\description{
Looks for corresponding FCL codes in country-specific
mapping tables from MDB files
}

