% Generated by roxygen2 (4.0.1): do not edit by hand
\name{imputeUnitValue}
\alias{imputeUnitValue}
\title{This function impute missing unit value}
\usage{
imputeUnitValue(data, unitValue, mirrorUnitValue)
}
\arguments{
\item{data}{The data frame}

\item{unitValue}{The unit value calculated for the reporting country.}

\item{mirrorUnitValue}{The unit value calculated for the partner country.}
}
\description{
Missing unit value are first imputed by it's mirror value,
followed by the median of the value.
}

