% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.SWSEnvir.R
\name{is.SWSEnvir}
\alias{is.SWSEnvir}
\title{Function to check is it SWS environment}
\usage{
is.SWSEnvir(varToCheck = "swsContext.datasets")
}
\arguments{
\item{varToCheck}{Variable to check.}
}
\value{
logical value. TRUE if there
}
\description{
Detecting existence of SWS Envir in calling env
by looking for swsContext.datasets
}
