% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_utilities.R
\name{list_rprt_dirs}
\alias{list_rprt_dirs}
\title{Provides list of directories with names matching reporting scheme.}
\usage{
list_rprt_dirs(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where report directories are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\value{
Character vector with full names of suitable directories.
}
\description{
Provides list of directories with names matching reporting scheme.
}
