% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livestockWeightsTable.R
\name{livestockWeightsTable}
\alias{livestockWeightsTable}
\title{Livestock weight table.}
\usage{
livestockWeightsTable()
}
\value{
Livestock weight table.
}
\description{
Livestock weight table.
}
