% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m49}
\alias{m49}
\title{Table of M49 area codes. Taken from
http://comtrade.un.org/db/mr/rfReportersList.aspx and 652 added manually.}
\format{An object of class \code{data.frame} with 295 rows and 10 columns.}
\usage{
m49
}
\description{
Table of M49 area codes. Taken from
http://comtrade.un.org/db/mr/rfReportersList.aspx and 652 added manually.
}
\keyword{datasets}

