% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m49faomap}
\alias{m49faomap}
\title{Comtrade M49 <-> FAO area list map.
This table should be used for reporters codes.}
\format{An object of class \code{data.frame} with 294 rows and 2 columns.}
\usage{
m49faomap
}
\description{
Comtrade M49 <-> FAO area list map.
This table should be used for reporters codes.
}
\keyword{datasets}

