% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapHS6toCPC.R
\name{mapHS6toCPC}
\alias{mapHS6toCPC}
\title{Map HS to CPC based on HS codes of length 6}
\usage{
mapHS6toCPC(tradedata, hs6maptable)
}
\arguments{
\item{tradedata}{Trade data frame.}

\item{hs6maptable}{Map table produced by \code{extract_hs6cpcmap} function.}
}
\value{
Data frame with unique reporter/flow/hs->cpc links.
}
\description{
Map HS to CPC based on HS codes of length 6
}
