% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapHS6toFCL.R
\name{mapHS6toFCL}
\alias{mapHS6toFCL}
\title{Map HS to FCL based on HS codes of length 6}
\usage{
mapHS6toFCL(tradedata, hs6maptable)
}
\arguments{
\item{tradedata}{Trade data frame.}

\item{hs6maptable}{Map table produced by \code{extract_hs6fclmap} function.}
}
\value{
Data frame with unique reporter/flow/hs->fcl links.
}
\description{
Map HS to FCL based on HS codes of length 6
}
