% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdbfaoareamap}
\alias{mdbfaoareamap}
\title{Mapping between character codes of areas in folders' names with
Jellyfish MDB files and FAO area codes}
\format{A data frame:
mdbarea - character vector with area codes in folders with MDB
faoarea - integer vector with FAO area codes}
\usage{
mdbfaoareamap
}
\description{
Produced by Onno
}
\keyword{datasets}
