% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltTradeData.R
\name{meltTradeData}
\alias{meltTradeData}
\title{Melt trade database}
\usage{
meltTradeData(data = NA)
}
\arguments{
\item{data}{data from SWS.}
}
\value{
Melted data.
}
\description{
Melt trade database
}
