% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorNonReporters.R
\name{mirrorNonReporters}
\alias{mirrorNonReporters}
\title{Mirror non-reporting and incomplete countries.}
\usage{
mirrorNonReporters(tradedata = NA, mirror = NA)
}
\arguments{
\item{tradedata}{Trade data.}

\item{mirror}{data.frame containing the code of the area, the
flow and the number of flows.}
}
\value{
\code{tradedata} with mirrored data for non-reporters added.
}
\description{
Use mirror trade flows for imputing trade data for non-reporting
countries and for countries that do not report a flow.
}
\details{
For non-reporting and incomplete countries exports of reporters will
become their imports, and vice versa. In order to take into account
the different international shipping costs, a CIF (Cost, Insurance,
Freight) / FOB (Free On Board) correction is done. Currently, the
CIF/FOB correction is set at 12% (imports are increased by 12% the
values of reporters' exports).
}
