% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorRules.R
\name{mirrorRules}
\alias{mirrorRules}
\title{Use mirror rules for balancing trade.}
\usage{
mirrorRules(data = NA, variable = "qty", suffix = "_m",
  flag = "flag_qty", group_rank = "accu_group", flow = "flow",
  main_flow = 1, discrepancy = "discrep_mirr", official_flags = c("X-",
  "-", "-h"))
}
\arguments{
\item{data}{data.}
}
\value{
Reshaped data with mirror flows, if necessary.
}
\description{
Use mirror rules for balancing trade.
}
