% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mirrorTrade}
\alias{mirrorTrade}
\title{This function mirrors trade by replacing missing trade if
corresponding trade is available and reported by the trading
partner.}
\usage{
mirrorTrade(data, reportingCountry, partnerCountry, reverseTradePrefix,
  valueColumns, flagColumns)
}
\arguments{
\item{data}{The data frame}

\item{reportingCountry}{The column represents the reporting country.}

\item{partnerCountry}{The column representing the partner country
of the trade.}

\item{reverseTradePrefix}{The prefix which indicates the flow is
reported by the partner country.}

\item{valueColumns}{The column which represents the traded value
and quantities.}

\item{flagColumns}{The column which represents the flag of the
trade.}
}
\description{
This function mirrors trade by replacing missing trade if
corresponding trade is available and reported by the trading
partner.
}

