% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movav.R
\name{movav}
\alias{movav}
\title{Moving average}
\usage{
movav(x, pkg = "native", mode = "centered", n = 5, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{pkg}{string: "native" (default) or the name of a package with
a function that computes moving averages. Besides "native", only
"zoo" is allowed (\code{zoo::rollapply} will be used).
respectively.}

\item{mode}{String: "centered" (default) indicates that the moving
average should be centered.}
}
\value{
A vector with the moving average of the input.
}
\description{
Compute the moving average of a vector.
}
\details{
The resulting moving average will be of order three.
}
