% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_area.R
\name{name_area}
\alias{name_area}
\title{A function to provide names for codes from different area classifications.}
\usage{
name_area(areacode, code_class = NULL)
}
\arguments{
\item{areacode}{Vector of area list codes.}

\item{code_class}{A character string indicating classification of area codes:
"fao", "m49" or "geonom".}
}
\value{
Character vector of area names.
}
\description{
A function to provide names for codes from different area classifications.
}
