% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preAggregateMultipleTLRows.R
\name{preAggregateMultipleTLRows}
\alias{preAggregateMultipleTLRows}
\title{Aggregate multiple rows in raw Tariff Line data.}
\usage{
preAggregateMultipleTLRows(rawdata = NA)
}
\arguments{
\item{rawdata}{Raw TL (a tibble).}
}
\value{
The original TL data after aggregation, when possible (see
  details).
}
\description{
Do a first aggregation of raw Tariff Line data, when possible.
}
\details{
Tariff Line (TL) data can have multiple rows. They can be aggregated
(by taking also into account the possible different units) when value,
quantity, and weight are present. Other two possibilities is to do the
aggregation when weight is always missing, but quantity is always present
and vice versa. When aggregation is not possible the record will be left
untouched as the missing variable will be imputed in the imputation step.
}
