% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMDBTable.R
\name{readMDBTable}
\alias{readMDBTable}
\title{Reads a specific table from an MDB file, prepared for
Jellyfish software.}
\usage{
readMDBTable(area, year, table, mdbdir = c("", "mnt", "essdata", "TradeSys",
  "TradeSys", "Countries"), colClasses = "character")
}
\description{
It is expected folder name and file name are the same and
are formed from country name abbreviation and year, for example,
USA_2011.
}

