% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_cpu_cores.R
\name{register_cpu_cores}
\alias{register_cpu_cores}
\title{Check for parallel support and register CPU cores}
\usage{
register_cpu_cores()
}
\value{
TRUE if CPU cores were registered
}
\description{
Check for parallel support and register CPU cores
}
