% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeInconsistentQuantityValue}
\alias{removeInconsistentQuantityValue}
\title{This function removes quantity and value which are inconsistent}
\usage{
removeInconsistentQuantityValue(data, quantity, value)
}
\arguments{
\item{data}{The data frame}

\item{quantity}{The column representing quantity.}

\item{value}{The column representing value}
}
\description{
Inconsistent value is defined as either one of the quantity or
value is zero while the respective is not. In this case, the zero
value will be replaced will missing value.
}

