% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeInvalidReporters.R
\name{removeInvalidReporters}
\alias{removeInvalidReporters}
\title{Remove invalid reporters.}
\usage{
removeInvalidReporters(tradedata)
}
\arguments{
\item{tradedata}{TL or ES trade data.}
}
\value{
TL or ES data without invalid reporters.
}
\description{
Remove invalid reporters.
}
