% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNonNumeric.R
\name{removeNonNumeric}
\alias{removeNonNumeric}
\title{Remove non numeric reporter/partner/hs.}
\usage{
removeNonNumeric(tradedata)
}
\arguments{
\item{tradedata}{TL or ES trade data.}
}
\value{
TL or ES data without nonnumeric reporter/partner/hs.
}
\description{
Remove non numeric reporter/partner/hs.
}
