% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeSelfTrade}
\alias{removeSelfTrade}
\title{This function removes self-trade, which is deemed to be erroneous}
\usage{
removeSelfTrade(data, reportingCountry, partnerCountry)
}
\arguments{
\item{data}{The data frame}

\item{reportingCountry}{The column which represents the reporting country.}

\item{partnerCounytry}{The column which represents the partner
country of the trade.}
}
\description{
This function removes self-trade, which is deemed to be erroneous
}

