% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportdirectory.R
\name{reportdirectory}
\alias{reportdirectory}
\title{Create directory for writing a report and supplementary files}
\usage{
reportdirectory(sws_user, year, build_id, create = TRUE, browsedir = TRUE)
}
\arguments{
\item{year}{Year to include in the name.}

\item{build_id}{Identificator to include in the name.}

\item{create}{Should the directory be created. TRUE by default.}

\item{browsedir}{Should the directory be opened in system-default file
browser. TRUE by default.}

\item{user}{User name to include in the name.}
}
\value{
String with full name of report directory.
}
\description{
Create directory for writing a report and supplementary files
}
