% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt.R
\name{rprt}
\alias{rprt}
\title{Function to call reporting instructions}
\usage{
rprt(data = NULL, report = NULL, ...)
}
\arguments{
\item{data}{Data set to make report on.}

\item{report}{Name of reporting function to call as string. Prefix "rprt_"
can be omitted.}

\item{...}{Additional arguments passed onto report function.}
}
\value{
Original data invisibly.
}
\description{
We want the trade module to run even one of reporting functions fails, so
every call of any reporting function should be wrapped by try(). Also we do
not want to export too many functions in the user space. So this function
could be the one from reporting functions to be exported.
}
