% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_fulltable.R
\name{rprt_fulltable}
\alias{rprt_fulltable}
\title{Adds full table to report}
\usage{
rprt_fulltable(dataset, level = "info", prefix = NULL, pretty_prop = TRUE,
  area_code_class = NULL, area_columns = c("reporter", "partner"))
}
\arguments{
\item{dataset}{Data frame to report}

\item{level}{Character of length one. Futile logger level to send messages
to. By default "info".}

\item{prefix}{Character of length one. Prefix for name to use in the report.
NULL by default.}

\item{pretty_prop}{Logical of length one. If TRUE all numeric columns with
names end with "_prop" will be converted for easy reading with
scales::percent.}

\item{area_code_class}{Character "fao", "m49", "geonom" or NULL. If not NULL
columns with area names will be added. NULL by default. See
?add_area_names for details.}

\item{area_columns}{Character vector with names of columns containing area
codes. By default c("reporter", "partner").}
}
\description{
Adds full table to report
}
