% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_glimpse0.R
\name{rprt_glimpse0}
\alias{rprt_glimpse0}
\title{Drops dplyr::glimpse's invisible value}
\usage{
rprt_glimpse0(tbl)
}
\arguments{
\item{tbl}{Data.}
}
\description{
It is required to include glimpse's output into
logging messages of futile.logger.
}
