% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hs2cpc_fulldata.R
\name{rprt_hs2cpc_fulldata}
\alias{rprt_hs2cpc_fulldata}
\title{Reporting on multiple matches during HS->CPC matching by hsInRange function.}
\usage{
rprt_hs2cpc_fulldata(tradedata, tradedataname = NULL)
}
\arguments{
\item{tradedata}{Trade data set with cpc column (after mapping process).}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
The function generates the following statistics per reporter, hs6,
reporter/hs6:
\itemize{
  \item Number of nonmatched records.
  \item Proportion of nonmatched records.
  \item Total value (cost) of nonmatched records.
  \item Proportion of total value of nonmatched records.}
}
