% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hs2cpc_multilinks.R
\name{rprt_hs2cpc_multilinks}
\alias{rprt_hs2cpc_multilinks}
\title{Reporting on multiple matches during HS->CPC matching by hsInRange function.}
\usage{
rprt_hs2cpc_multilinks(uniqhs, tradedataname = NULL)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow, hsext and cpc
(produced in mapHS2CPC()).}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
Reporting on multiple matches during HS->CPC matching by hsInRange function.
}
