% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hs2cpc_nolinks.R
\name{rprt_hs2cpc_nolinks}
\alias{rprt_hs2cpc_nolinks}
\title{Reporting on results of HS ranges matches by function hsInRange.}
\usage{
rprt_hs2cpc_nolinks(uniqhs, tradedataname = NULL)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow, hsext and cpc
(produced in mapHS2CPC()).}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
Reporting on results of HS ranges matches by function hsInRange.
}
