% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hs2fcl_multilinks.R
\name{rprt_hs2fcl_multilinks}
\alias{rprt_hs2fcl_multilinks}
\title{Reporting on multiple matches during HS->FCL matching by hsInRange function.}
\usage{
rprt_hs2fcl_multilinks(uniqhs, tradedataname = NULL)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow, hsext and fcl.
(produced in mapHS2FCL()).}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
Reporting on multiple matches during HS->FCL matching by hsInRange function.
}
