% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hs6cpcmap.R
\name{rprt_hs6cpcmap}
\alias{rprt_hs6cpcmap}
\title{Reports on results of extraction of HS6->CPC map}
\usage{
rprt_hs6cpcmap(hs6cpcmap = NULL)
}
\arguments{
\item{hs6cpcmap}{produced by extract_hs6cpcmap function}
}
\description{
Reports on results of extraction of HS6->CPC map
}
