% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hs6fcl_results.R
\name{rprt_hs6fcl_results}
\alias{rprt_hs6fcl_results}
\title{Reports on results of HS6->FCL mapping process.}
\usage{
rprt_hs6fcl_results(data = NULL, tradedataname = NULL)
}
\arguments{
\item{data}{Data.}

\item{tradedataname}{Trade data name.}
}
\description{
Reports on results of HS6->FCL mapping process.
}
