% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hs6fclmap.R
\name{rprt_hs6fclmap}
\alias{rprt_hs6fclmap}
\title{Reports on results of extraction of HS6->FCL map}
\usage{
rprt_hs6fclmap(data = NULL)
}
\arguments{
\item{data}{produced by extract_hs6fclmap function}
}
\description{
Reports on results of extraction of HS6->FCL map
}
