% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hschanged.R
\name{rprt_hschanged}
\alias{rprt_hschanged}
\title{Reporting on HS code changes in trade data during HS->FCL mapping
process.}
\usage{
rprt_hschanged(uniqhs, tradedataname = NULL)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow, maxhslength, hsextchar
and hsext. (produced in mapHS2FCL()).}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
Reporting on HS code changes in trade data during HS->FCL mapping
process.
}
