% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hsfclmap.R
\name{rprt_hsfclmap}
\alias{rprt_hsfclmap}
\title{Function to produce report on HS->FCL mapping table}
\usage{
rprt_hsfclmap(maptable, year)
}
\arguments{
\item{maptable}{Data frame with hsfclmap table.}

\item{year}{Integer. Reporting year.}
}
\value{
TBD
}
\description{
Function calculates:
}
\details{
\itemize{
  \item Number of mapping records per country.
  \item Length of HS codes.
  \item Odd/even digit length of HS codes.
}
}
\seealso{
See \url{https://github.com/SWS-Methodology/faoswsTrade/issues/83}
  for problem background.
}
