% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hsinrange.R
\name{rprt_hsinrange}
\alias{rprt_hsinrange}
\title{Reporting on results of HS ranges matches by function hsInRange.}
\usage{
rprt_hsinrange(uniqhs, tradedataname = NULL)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow, hsext and fcl.
(produced in mapHS2FCL()).}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
Reporting on results of HS ranges matches by function hsInRange.
}
