% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_hslength.R
\name{rprt_hslength}
\alias{rprt_hslength}
\title{Reporting on reporter/flow/hs unique combinations during HS->FCL mapping
process.}
\usage{
rprt_hslength(hslength, tradedataname = NULL)
}
\arguments{
\item{hslength}{Data frame with columns reporter, flow and maxhslength
(produced by maxHSLength().}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics. Currently hslength itself.
}
\description{
Reporting on reporter/flow/hs unique combinations during HS->FCL mapping
process.
}
