% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_map_hschanged.R
\name{rprt_map_hschanged}
\alias{rprt_map_hschanged}
\title{Reporting on modifications of HS codes in mapping table during HS->FCL mapping
process.}
\usage{
rprt_map_hschanged(maptable, tradedataname = NULL)
}
\arguments{
\item{maptable}{Data frame with columns reporter, flow, maxhslength, fromcodeextchar
and fromcodeext. (produced in mapHS2FCL()).}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
Reporting on modifications of HS codes in mapping table during HS->FCL mapping
process.
}
