% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_rawtradedata.R
\name{rprt_rawtradedata}
\alias{rprt_rawtradedata}
\title{Report table on raw trade data}
\usage{
rprt_rawtradedata(tradedata = NULL, tradedataname = NULL)
}
\arguments{
\item{tradedata}{Trade data.}

\item{tradedataname}{Trade data name.}
}
\description{
Report table on raw trade data
}
