% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_report.R
\name{rprt_report}
\alias{rprt_report}
\title{Adds info to futile.logger report}
\usage{
rprt_report(msg, level = "info")
}
\arguments{
\item{msg}{Object to add}

\item{level}{Character. Futile logger level to use. INFO by default.}
}
\description{
Adds info to futile.logger report
}
