% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_uniqhs.R
\name{rprt_uniqhs}
\alias{rprt_uniqhs}
\title{Reporting on reporter/flow/hs unique combinations during HS->FCL mapping
process.}
\usage{
rprt_uniqhs(uniqhs, tradedataname = NULL)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow and hs.}

\item{tradedataname}{Character of length 1. Most likely `esdata` or `tldata`.}
}
\value{
Data frame with summary statistics.
}
\description{
Reporting on reporter/flow/hs unique combinations during HS->FCL mapping
process.
}
