% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFileForPPR.R
\name{saveFileForPPR}
\alias{saveFileForPPR}
\title{Save file for pre-processing report.}
\usage{
saveFileForPPR(data = NA, file = NA, type = NA, sleep = 10,
  times = 10)
}
\arguments{
\item{data}{Data to save.}

\item{file}{File name.}

\item{type}{FIle type: can be either "csv" or "rds" and the data will
be saved as a csv or rds file, respectively. If this parameter is
missing, the file type will be read from the extension of the `file`
parameter (and will be used only if it is "csv" or "rds").}

\item{sleep}{Amount of time to sleep between each one of the `times`
attempted saves.}

\item{times}{How many times try to save the file.}
}
\value{
Nothing: used for side effect (save file).
}
\description{
Save file for pre-processing report.
}
