% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel1CPC.R
\name{sel1CPC}
\alias{sel1CPC}
\title{Select one record from multiple matches during HS->CPC mapping process.}
\usage{
sel1CPC(hscpcmatch, maptable, cur_yr = NULL)
}
\arguments{
\item{hscpcmatch}{Data frame with columns reporter, flow, hs, hsext, cpc,
datumid, maplinkid}

\item{maptable}{Data frame with HS->CPC mapping table.}

\item{cur_yr}{Integer, year the trade module run on.}
}
\value{
Data frame similar to hscpcmatch but with all duplicates removed.
}
\description{
Select one record from multiple matches during HS->CPC mapping process.
}
