% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel1FCL.R
\name{sel1FCL}
\alias{sel1FCL}
\title{Select one record from multiple matches during HS->FCL mapping process.}
\usage{
sel1FCL(hsfclmatch, maptable)
}
\arguments{
\item{mapsubset}{Data frame with columns reporter, flow, hs, fcl}
}
\description{
Select one record from multiple matches during HS->FCL mapping process.
}
