% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sws_dev_settings.R
\name{set_sws_dev_settings}
\alias{set_sws_dev_settings}
\title{Set SWS environment in development mode}
\usage{
set_sws_dev_settings(localsettingspath = NULL)
}
\arguments{
\item{localsettingspath}{Path to a file with local SWS settings. See
faoswsModules::ReadSettings for details.}
}
\value{
invisible settings.
}
\description{
Massive side effects in R session and system environment.
}
