% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFlag2.R
\name{setFlag2}
\alias{setFlag2}
\title{Set flags in trade data.}
\usage{
setFlag2(data = NA, condition = NA, type = NA, flag = NA,
  variable = NA)
}
\arguments{
\item{type}{The type of flag to set: "status" or "method".}

\item{flag}{The value of the flag to set (e.g., "e" for estimated data).}

\item{variable}{The variable that gets the flag: "value", "quantity".}
}
\value{
Populates the flag variable.
}
\description{
The generated flag will be a numeric variable like 1XY where
X and Y are dummies that are equal to 1 if the flag applies to
the \code{value} or \code{quantity} variables, respectively, and
0 otherwise.
}
