% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFlag3.R
\name{setFlag3}
\alias{setFlag3}
\title{Set flags in trade data.}
\usage{
setFlag3(data = NA, condition = NA, type = NA, flag = NA,
  variable = NA)
}
\arguments{
\item{data}{Trade data.}

\item{condition}{A logical condition where the flags should be changed.}

\item{type}{The type of flag to set: "status" or "method".}

\item{flag}{The value of the flag to set (e.g., "e" for estimated data).}

\item{variable}{The variable that gets the flag: "value", "weight",
"quantity".}
}
\value{
Populates the flag variable.
}
\description{
The generated flag will be a numeric variable like 1XYZ where
X, Y and Z are dummies that are equal to 1 if the flag applies to
the \code{value} or \code{weight} or \code{quantity} variables,
respectively, and 0 otherwise.
}
