% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificCorrectionsHS.R
\name{specificCorrectionsHS}
\alias{specificCorrectionsHS}
\title{Specific corrections to HS data.}
\usage{
specificCorrectionsHS(tradedata)
}
\arguments{
\item{data}{data.}
}
\value{
Data with corrections applied.
}
\description{
The function will adjust weight and quantity for given HS codes.
}
\details{
NOTE THAT IT IS CURRENTLY SPECIFIC FOR A GIVEN REPORTER AND YEAR.
}
