% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFlags.R
\name{sumFlags}
\alias{sumFlags}
\title{Sum the flags indicator variables}
\usage{
sumFlags(new = NA, old = NA, flags = NA)
}
\arguments{
\item{new}{The new flag. Should be a scalar.}

\item{old}{The old flags. Can be a vector.}

\item{flags}{A numeric vector of flags to be summed.}
}
\value{
An indicator variable.
}
\description{
This function will sum two indicator variables of the type
XYZ where X, Y or Z are 0 or 1, and the result will be itself
a XYZ variable. Note: if \code{new} and \code{old} are used,
\code{flags} should not be used, and vice versa.
}
