% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapFlags.R
\name{swapFlags}
\alias{swapFlags}
\title{Change position of flags.}
\usage{
swapFlags(variable = stop("'variable' is required."),
  swap = stop("'swap' is required."), condition = NULL)
}
\description{
Change position of flags.
}
