% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfvalid}
\alias{tfvalid}
\title{Valid dataset from database to be compared with new procedure}
\format{A data frame with seven columns as integers: year, reporter,
 partner, fcl, quantity and value}
\usage{
tfvalid
}
\description{
Taken as CSV from Giorgio Sorbara Giorgio.Sorbara@fao.org
}
\keyword{datasets}
