% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailingDigits.R
\name{trailingDigits}
\alias{trailingDigits}
\title{Adds zeros to the end of hs column to make both columns of equal length}
\usage{
trailingDigits(from, to, digit)
}
\description{
Adds zeros to the end of hs column to make both columns of equal length
}

