% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailingDigits2.R
\name{trailingDigits2}
\alias{trailingDigits2}
\title{Other version of adding of trailing digits. We should check it.}
\usage{
trailingDigits2(client, maxlength, digit)
}
\description{
Other version of adding of trailing digits. We should check it.
}

