% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_all_reports.R
\name{ts_all_reports}
\alias{ts_all_reports}
\title{General function to produce all time-series reports}
\usage{
ts_all_reports(collection_path = NULL, prefix = NULL,
  ts_reports_path = NULL, complete = TRUE)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}

\item{ts_reports_path}{String with path where time series reports should be
saved. If NULL (default) the directory with TS reports will be saved where
single-year reports are located (collection_path).}

\item{complete}{Logical If TRUE (default) all reports (raw data, mapping,
etc.) will be generated, if FALSE only raw data reports will be produced.}
}
\value{
NULL inivisibly.
}
\description{
General function to produce all time-series reports
}
\examples{
ts_all_reports(Sys.getenv("HOME"), "complete_tf_cpc")
}
