% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_corrections_not_applied.R
\name{ts_corrections_not_applied}
\alias{ts_corrections_not_applied}
\title{Function to produce the list of corrections not applied.}
\usage{
ts_corrections_not_applied(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\value{
A table with corrections that were not applied.
}
\description{
Function to produce the list of corrections not applied.
}
\examples{
ts_corrections_not_applied("/mnt/storage/sws_share/sas", "complete_tf_cpc")

}
