% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_count_records.R
\name{ts_count_records}
\alias{ts_count_records}
\title{Count number of records across years.}
\usage{
ts_count_records(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\description{
Table with the number of records by reporter/year/flow.
It also contains the maximum HS length, the percentage
difference in trade flows per year and whether the
maximum HS length changed with respect to the previous year.
}
