% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_flows_check.R
\name{ts_flows_check}
\alias{ts_flows_check}
\title{Check if value and quantity and weight exist.}
\usage{
ts_flows_check(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\value{
NULL invisibly.
}
\description{
Check if value and quantity and weight exist.
}
\examples{
ts_content_check("/mnt/storage/sws_share/sas", "complete_tf_cpc")

}
