% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fulldata_hsfcl_nolinks.R
\name{ts_fulldata_hsfcl_nolinks}
\alias{ts_fulldata_hsfcl_nolinks}
\title{Extract all unmapped HS codes across years}
\usage{
ts_fulldata_hsfcl_nolinks(collection_path = NULL, prefix = NULL)
}
\description{
Extract all unmapped HS codes across years
}
