% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_hsfcl_nolinks_statistic.R
\name{ts_hsfcl_nolinks_statistic}
\alias{ts_hsfcl_nolinks_statistic}
\title{Extracts time series of not mapped HS codes number}
\usage{
ts_hsfcl_nolinks_statistic(collection_path = NULL, prefix = NULL)
}
\value{
Data frame
}
\description{
Extracts time series of not mapped HS codes number
}
\examples{
library(ggplot2)
library(dygraph)
library(dplyr)

ts1 <- ts_hsfcl_nolinks_statistic("/mnt/storage/sws_share/chr", "complete_tf_cpc")

ts1 \%>\%
  group_by(year) \%>\%
  summarize(total_nolinks = sum(nolinks)) \%>\%
  ungroup() \%>\%
  ggplot(aes(year, total_nolinks)) +
  geom_bar(stat = "identity")

ts1 \%>\%
  group_by(year) \%>\%
  summarize(total_nolinks = sum(nolinks)) \%>\%
  ungroup() \%>\%
  dygraph(
    ylab = "Total number of unmapped HS (duplicates across years possible)",
    main = "Unmapped HS codes (Tariff line and Eurostat)")
}
