% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_missing_records.R
\name{ts_missing_records}
\alias{ts_missing_records}
\title{Report the number of records with missing quantity/weight.}
\usage{
ts_missing_records(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\description{
Report the number of records with missing quantity/weight.
}
