% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_nolinks_data.R
\name{ts_nolinks_data}
\alias{ts_nolinks_data}
\title{Extracts time series of not mapped HS codes number}
\usage{
ts_nolinks_data(collection_path = NULL, prefix = NULL)
}
\description{
Extracts time series of not mapped HS codes number
}
\examples{
library(ggplot2)
library(dygraph)
library(dplyr)

ts1 <- ts_nolinks_data("/mnt/storage/sws_share/chr", "complete_tf_cpc")

ts1 \%>\%
  group_by(year) \%>\%
  summarize(total_nolinks = sum(nolinks)) \%>\%
  ungroup() \%>\%
  ggplot(aes(year, total_nolinks)) +
  geom_line() +
  scale_y_log10()

ts1 \%>\%
  group_by(year) \%>\%
  summarize(total_nolinks = sum(nolinks)) \%>\%
  ungroup() \%>\%
  dygraph(
    ylab = "Total number of unmapped HS (duplicates across years possible)",
    main = "Unmapped HS codes (Tariff line and Eurostat)")
}
