% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_non_reporters.R
\name{ts_non_reporters}
\alias{ts_non_reporters}
\title{Function to produce the list of non-reporting countries.}
\usage{
ts_non_reporters(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\value{
A table with countries/years where the cell is 9 if the country
  does not report completely for that year, 1 if it does not report
  imports, 2 if it does not report exports.
}
\description{
Function to produce the list of non-reporting countries.
}
\examples{
ts_non_reporters("/mnt/storage/sws_share/sas", "complete_tf_cpc")

}
