% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preproc.R
\name{ts_preproc}
\alias{ts_preproc}
\title{Extracts time series data from preprocessing reports}
\usage{
ts_preproc(collection_path = NULL, prefix = NULL)
}
\description{
Extracts time series data from preprocessing reports
}
