% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reporters.R
\name{ts_reporters}
\alias{ts_reporters}
\title{Function to produce the list of reporters cointained in the file.}
\usage{
ts_reporters(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\value{
NULL invisibly.
}
\description{
Function to produce the list of reporters cointained in the file.
}
\examples{
ts_reporters("/mnt/storage/sws_share/sas", "complete_tf_cpc")

}
