% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_utilities.R
\name{ts_write_rprt}
\alias{ts_write_rprt}
\title{Writes table with time series report to specific directory as csv file.}
\usage{
ts_write_rprt(rprt = NULL, ts_reports_path = NULL, rprt_name = NULL)
}
\arguments{
\item{rprt}{Data frame to save.}

\item{ts_reports_path}{Directory path to save csv file.}

\item{rprt_name}{Name of report.}
}
\description{
Writes table with time series report to specific directory as csv file.
}
