% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unsdpartnersblocks}
\alias{unsdpartnersblocks}
\title{UNSD partner blocksreference list}
\format{A data frame}
\usage{
unsdpartnersblocks
}
\description{
Taken from Melissa Paca paca@un.org
}
\keyword{datasets}
