% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInfoTable.R
\name{updateInfoTable}
\alias{updateInfoTable}
\title{Update the table with information of module runs.}
\usage{
updateInfoTable(year = NA, mode = NA, results = NA)
}
\arguments{
\item{year}{The year.}

\item{mode}{String: which type of update will be done. "restart" means
only the "restarted" date string will be updated (it will also update
the "info" variable by setting it to "restarted"; "save" will update the
table with information to the data inserted/appended/ignored/discarded,
and will update the "saved" variable (it will also update the "info"
variable by inserting the hours required by the module).}

\item{results}{The object returned by faosws::SaveData().}
}
\value{
NULL.
}
\description{
Update the table with information of module runs.
}
