% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateTradeQuantity}
\alias{updateTradeQuantity}
\title{This function calculates the trade quantity based on trade value
and trade unit value.}
\usage{
updateTradeQuantity(data, unitValue, value, quantity)
}
\arguments{
\item{data}{The data table}

\item{unitValue}{The column representing trade unit value.}

\item{value}{The column representing trade value.}

\item{quantity}{The column representing trade quantity.}
}
\description{
This function calculates the trade quantity based on trade value
and trade unit value.
}

