% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateTradeValue}
\alias{updateTradeValue}
\title{This function updates the trade value.}
\usage{
updateTradeValue(data, unitValue, value, quantity)
}
\arguments{
\item{data}{The data table}

\item{unitValue}{The column representing trade unit value.}

\item{value}{The column representing trade value.}

\item{quantity}{The column representing trade quantity.}
}
\description{
This function updates the trade value.
}

