% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useAdjustments.R
\name{useAdjustments}
\alias{useAdjustments}
\title{Use specific adjustment factors for trade data.}
\usage{
useAdjustments(tradedata = stop("\\"tradedata\\" is missing."),
  year = stop("\\"year\\" is missing."),
  adjustments = stop("\\"adjustments\\" is missing."), parallel = FALSE,
  dbg = faosws::CheckDebug(), PID = NA, verbose = TRUE)
}
\arguments{
\item{tradedata}{Trade data.}

\item{year}{Numeric.}

\item{adjustments}{Adjustment factors.}

\item{parallel}{Logical: FALSE (dafault), no multicore; TRUE, multicore.}

\item{dbg}{Logical: see \code{\link[faosws]{CheckDebug}}.}

\item{PID}{Numeric: a process id (can also be a string). Used if
\code{verbose == TRUE}.}

\item{verbose}{Logical: TRUE (default), message shown; FALSE, no message.}
}
\value{
The same dataset after the application of adjustment factors.
}
\description{
Use specific adjustment factors for trade data.
}
