% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useValidationCorrections.R
\name{useValidationCorrections}
\alias{useValidationCorrections}
\title{Use the corrections deriving from validation.}
\usage{
useValidationCorrections(data, corrections)
}
\arguments{
\item{data}{complete_trade_flow_cpc dataframe.}

\item{corrections}{Corrections table.}
}
\value{
A list with a dataframe of corrections to drop and a
  subset of complete_trade_flow_cpc with corrections applied.
}
\description{
Use the corrections deriving from validation.
}
