% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validation}
\alias{validation}
\title{This is the wrapper function of the mirror validation and range validation.}
\usage{
validation(data, value, mirrorValue, ratioBoundary = 3, plot = FALSE)
}
\arguments{
\item{data}{The data frame}

\item{value}{The value reported by the reporting country}

\item{mirrorValue}{The value reported by the partner country}

\item{ratioBoundary}{A real number which is multiplied to the
slope of the total least squares to determine the decision bounds.}

\item{plot}{logical, whether the graph of the TLS should be plotted.}
}
\description{
This is the wrapper function of the mirror validation and range validation.
}

