% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validationByMirrorValue}
\alias{validationByMirrorValue}
\title{This function validates trade by compare the unit value of the
reporting country and the partner country.}
\usage{
validationByMirrorValue(value, mirrorValue, ratioBoundary, plot = FALSE)
}
\arguments{
\item{value}{The value reported by the reporting country}

\item{mirrorValue}{The value reported by the partner country.}

\item{ratioBoundary}{A real number which is multiplied to the
slope of the total least squares to determine the decision bounds.}
}
\description{
The validation is done via total least squares, value outside of
the ratio bound are perceived as unreliable and will be replaced
by the expected value of the total least squares fit.
}

