% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validationByRange}
\alias{validationByRange}
\title{This function validates value by removing values which are outliers.}
\usage{
validationByRange(value)
}
\arguments{
\item{value}{The vector of value.}
}
\description{
Tukey's method is used to determine whether a value is an outlier.
}

