% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyTarifflinePartners.R
\name{verifyTarifflinePartners}
\alias{verifyTarifflinePartners}
\title{Converts reporter-specific codes of partners into FAO area codes}
\usage{
verifyTarifflinePartners(tldata, ctdata)
}
\arguments{
\item{Tariff-line}{data}

\item{Comtrade}{data}
}
\value{
Data frame with columns: Comtrade M49 reporter codes, Tariffline
reported partner codes and corresponding Comtrade M49 area codes
}
\description{
Converts reporter-specific codes of partners into FAO area codes
}

