% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_area_names.R
\name{add_area_names}
\alias{add_area_names}
\title{A function to add columns with area names to a data frame where reporter or
partner columns exist.}
\usage{
add_area_names(data, codeclass = "fao", areacolumns = c("reporter",
  "partner"))
}
\arguments{
\item{data}{A data frame where reporter or partner columns exists.}

\item{codeclass}{A character string indicating classification of area codes:
"fao", "m49" or "geonom". By default "fao".}

\item{areacolumns}{A character vector indicating names of columns with area
codes. By default c("reporter", "partner").}
}
\value{
Original data frame with added columns.
}
\description{
A function to add columns with area names to a data frame where reporter or
partner columns exist.
}
