% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_utilities.R
\name{extract_rprt_elem}
\alias{extract_rprt_elem}
\title{Extract an element from list stored in RDS archive in report directory.}
\usage{
extract_rprt_elem(collection_path = NULL, prefix = NULL, elem = NULL)
}
\arguments{
\item{collection_path}{String with path where report directories are stored.}

\item{prefix}{A string with report directory name prefix (without trailing
underscore. By default NULL.}

\item{elem}{Character vector with names of elements to extract.}
}
\value{
List of extracted tables
}
\description{
Extract an element from list stored in RDS archive in report directory.
}
