% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoAreaName.R
\name{faoAreaName}
\alias{faoAreaName}
\title{Assign area names to M49 or geonom codes.}
\usage{
faoAreaName(areacode, swsdomain = "faostat_one", swsdataset = "FS1_SUA_UPD",
  swsdimension = "geographicAreaFS")
}
\description{
Assign area names to M49 or geonom codes.
}
