% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6faointerest}
\alias{hs6faointerest}
\title{HS codes of length 6 from HSagric_filter.xls what take part in futher processing
of agricultural trade data.}
\format{A numeric vector of length 21352.}
\usage{
hs6faointerest
}
\description{
HS codes of length 6 from HSagric_filter.xls what take part in futher processing
of agricultural trade data.
}
\keyword{datasets}
