% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsInRange.R
\name{hsInRange}
\alias{hsInRange}
\title{Looks for corresponding CPC codes in country-specific
mapping tables from MDB files}
\usage{
hsInRange(uniqhs, maptable, parallel = FALSE)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow, hsext.}

\item{maptable}{Data frame with HS->CPC mapping with columns area,
flow, fromcode, tocode, cpc, fcl}

\item{parallel}{Logical. Should multicore backend be used.}
}
\value{
Data frame with columns reporter, flow, datumid, hs, hsext, cpc, fcl.
  datumid holds row numbers of original dataset. hs is input hs. hsext is
  input hs with additional zeros if requires. If there are multiple
  HS->CPC matchings, all of them are returned with similar id. If
  there were no matching CPC codes, NA in cpc column is returned.
}
\description{
Looks for corresponding CPC codes in country-specific
mapping tables from MDB files
}
