% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsInRange.R
\name{hsInRange}
\alias{hsInRange}
\title{Looks for corresponding FCL codes in country-specific
mapping tables from MDB files}
\usage{
hsInRange(uniqhs, maptable, parallel = FALSE)
}
\arguments{
\item{uniqhs}{Data frame with columns reporter, flow, hs.}

\item{maptable}{Data frame with HS->FCL mapping with columns area,
flow, fromcode, tocode, fcl}

\item{parallel}{Logical. Should multicore backend be used.}
}
\value{
Data frame with columns id, area, flow, hsorig, hsext, fcl.
  id holds row numbers of original dataset. hsorig is input hs. hsext is
  input hs with additional zeros if requires. If there are multiple
  HS->FCL matchings, all of them are returned with similar id. If
  there were no matching FCL codes, NA in fcl column is returned.
}
\description{
Looks for corresponding FCL codes in country-specific
mapping tables from MDB files
}
\details{
Input hs, areacode and flowname columns are used to build
  data frame with hs-codes to convert. Probably it is easier to
  pass a data frame, then separate vectors.
}
