% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapHS2FCL.R
\name{mapHS2FCL}
\alias{mapHS2FCL}
\title{Umbrella function to map HS codes to FCL codes}
\usage{
mapHS2FCL(tradedata, maptable, hs6maptable, parallel = FALSE)
}
\arguments{
\item{tradedata}{Trade data frame. Columns reporter, flow and hs are expected.}

\item{maptable}{Mapping table.}

\item{parallel}{Logical, should multicore backend be used. False by default.}
}
\value{
Data frame with unique combinations reporter/flow/hs/fcl.
}
\description{
The function takes trade data set with mapping table and finds corresponding
FCL codes for specific reporter, flow and HS code combinations.
}
