% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxHSLength.R
\name{maxHSLength}
\alias{maxHSLength}
\title{Prepares for numerical comparison vector of hs codes from
trade dataset with hs codes range from mapping table.}
\usage{
maxHSLength(uniqhs, mapdataset, parallel = FALSE)
}
\arguments{
\item{uniqhs}{Data frame with columns `reporter`, `flow` and `hs` containing
unique combinations from a trade data set.}

\item{mapdataset}{Data frame with mapping table containing
at least columns area, flow, fromcode, tocode}
}
\value{
A data frame with columns `reporter`, `flow` and `maxhslength`.
}
\description{
Prepares for numerical comparison vector of hs codes from
trade dataset with hs codes range from mapping table.
}
