% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_fulltable.R
\name{rprt_fulltable}
\alias{rprt_fulltable}
\title{Adds full table to report}
\usage{
rprt_fulltable(dataset, level = "info")
}
\arguments{
\item{dataset}{Data frame to report}
}
\description{
Adds full table to report
}
