% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprt_writetable.R
\name{rprt_writetable}
\alias{rprt_writetable}
\title{Writes full table for report collection}
\usage{
rprt_writetable(dataset, prefix = NULL)
}
\arguments{
\item{dataset}{Data frame to save}
}
\description{
Writes full table for report collection
}
