% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapFlags.R
\name{swapFlags}
\alias{swapFlags}
\title{Change position of flags.}
\usage{
swapFlags(variable = stop("'variable' is required."),
  swap = stop("'swap' is required."), condition = NULL)
}
\arguments{
\item{variable}{Varaible name.}

\item{swap}{How to swap flags (e.g., '\\2\\1\\3').}

\item{condition}{Logical condition that indicates which
flags will be swapped (when TRUE).}
}
\description{
Change position of flags.
}
