% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_non_reporters.R
\name{ts_non_reporters}
\alias{ts_non_reporters}
\title{Function to produce the list of non-reporting countries.}
\usage{
ts_non_reporters(collection_path = NULL, prefix = NULL)
}
\arguments{
\item{collection_path}{String with path where single-year report directories
are stored.}

\item{prefix}{String with report directory name prefix without trailing
underscore. By default NULL.}
}
\value{
NULL invisibly.
}
\description{
Function to produce the list of non-reporting countries.
}
\examples{
ts_non_reporters("/mnt/storage/sws_share/sas", "flows_to_mirror_raw")

}
